+++ mangadl/__init__.py
@@ -1,3 +1,3 @@
 """MangaDL - core package."""
 
-__version__ = "1.4.22"
+__version__ = "1.4.23"